+++ packages/ragbits-core/src/ragbits/core/embeddings/__init__.py
@@ -1,5 +1,16 @@
-from .base import Embeddings, EmbeddingsOptionsT, EmbeddingType
-from .litellm import LiteLLMEmbeddings
-from .noop import NoopEmbeddings
+from .base import Embedder, EmbedderOptionsT, EmbeddingType
+from .litellm import LiteLLMEmbedder
+from .noop import NoopEmbedder
+from .sparse import BagOfTokens, SparseEmbedder, SparseEmbedderOptionsT
 
-__all__ = ["EmbeddingType", "Embeddings", "EmbeddingsOptionsT", "LiteLLMEmbeddings", "NoopEmbeddings"]
+__all__ = [
+    "BagOfTokens",
+    "BagOfTokens",
+    "Embedder",
+    "EmbedderOptionsT",
+    "EmbeddingType",
+    "LiteLLMEmbedder",
+    "NoopEmbedder",
+    "SparseEmbedder",
+    "SparseEmbedderOptionsT",
+]