+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/providers/unstructured/images.py
@@ -7,7 +7,7 @@ from unstructured.documents.elements import Element as UnstructuredElement
 from unstructured.documents.elements import ElementType
 
 from ragbits.core.llms.base import LLM, LLMType
-from ragbits.core.llms.factory import get_default_llm
+from ragbits.core.llms.factory import get_preferred_llm
 from ragbits.core.prompt import Prompt
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
 from ragbits.document_search.documents.element import Element, ImageElement
@@ -93,7 +93,7 @@ class UnstructuredImageProvider(UnstructuredDefaultProvider):
         img_bytes = crop_and_convert_to_bytes(image, top_x, top_y, bottom_x, bottom_y)
         prompt = _ImagePrompt(_ImagePromptInput(images=[img_bytes]))
         if self.image_describer is None:
-            llm_to_use = self._llm if self._llm is not None else get_default_llm(LLMType.VISION)
+            llm_to_use = self._llm if self._llm is not None else get_preferred_llm(LLMType.VISION)
             self.image_describer = ImageDescriber(llm_to_use)
         image_description = await self.image_describer.get_image_description(prompt=prompt)
         return ImageElement(